Added type checking to improve code completion.
Added type stubs for all existing modules. This does not change the
source code of the modules, it just adds `.pyi` files of the same names.
The purpose is to make this information available to static type
checkers such as Mypy as well as to IDE tools for code completion. Since
the annotations are not inline, they are not available at run-time.

We are using BasedMypy instead of Mypy so that we have access to JPype's
type annotations, which Mypy ignores because a `py.typed` file is
missing from the installed package. We also prefer to not annotate
return types if they are just `None`, which is not possible with vanilla
Mypy. Unfortunately that also means that some (if not all) IDEs
currently misreport those return types as `Any`.