Split `dev` dependencies into `build`/`test`/`docs`.
It's sometimes nice to have smaller subsets of the optional dependencies
available, for example when building the documentation on Read-the-Docs.
It makes that job slightly faster. Though ultimately is not all that
important. The `dev` extra is the same as before and is what should be
installed for local development. We may be using the other subsets if
we (ever) add CI.